requirements.txt
numpy~=1.19.2
opencv-python~=4.4.0.44
mss~=6.0.0
pyserial~=3.4
scipy~=1.5.3